requirements.txt
playwright
playwright-stealth
google-genai
pillow
requests
python-dotenv
psycopg2-binary
torch
transformers
fastmcp
pgvector==0.2.5
fastapi